FIX security breach on user_id based requests

+++ app/controllers/devices_controller.rb
@@ -8,33 +8,39 @@ class DevicesController < ApplicationController
     end
 
     def show
-        devices = Device.where(user_id: params[:user_id]) 
-        if !devices.blank?
-            render json: { status: "SUCCESS", message: "All devices currently assigned to requested user", data: devices }, status: :ok
-        else
-            raise(ActiveRecord::RecordNotFound)
+        if check_current_user(params[:user_id])
+            devices = Device.where(user_id: params[:user_id]) 
+            if !devices.blank?
+                render json: { status: "SUCCESS", message: "All devices currently assigned to requested user", data: devices }, status: :ok
+            else
+                raise(ActiveRecord::RecordNotFound)
+            end
         end
     end
 
     def create
-        return unless check_checksum(params[:checksum].to_i, params[:device_id])
-        device = Device.new(device_params)
-        begin
-            device.save
-        rescue ActiveRecord::InvalidForeignKey => exception
-            render json: { status: 'ERROR', message: 'Erreur : le user-id n\'existe pas', data: exception.message }, status: :not_found
-        rescue ActiveRecord::RecordNotUnique => exception
-            render json: { status: 'ERROR', message: 'Erreur : ce Sensory Captor existe déjà', data: exception.message }, status: :unprocessable_entity
-        else
-            render json: { status: 'SUCCESS', message: "Sauvegarde réussie !", data: device }, status: :ok
+        if check_current_user(params[:user_id])
+            return unless check_checksum(params[:checksum].to_i, params[:device_id])
+            device = Device.new(device_params)
+            begin
+                device.save
+            rescue ActiveRecord::InvalidForeignKey => exception
+                render json: { status: 'ERROR', message: 'Erreur : le user-id n\'existe pas', data: exception.message }, status: :not_found
+            rescue ActiveRecord::RecordNotUnique => exception
+                render json: { status: 'ERROR', message: 'Erreur : ce Sensory Captor existe déjà', data: exception.message }, status: :unprocessable_entity
+            else
+                render json: { status: 'SUCCESS', message: "Sauvegarde réussie !", data: device }, status: :ok
+            end
         end
     end
 
     def update
         device = Device.find_by(device_id: params[:id])
         if device
-            device.update_attributes(device_params)
-            render json: { status: "SUCCESS", message: "Device updated", data: device }, status: :ok
+            if check_current_user(device.user_id)
+                device.update_attributes(device_params)
+                render json: { status: "SUCCESS", message: "Device updated", data: device }, status: :ok
+            end
         else
             render json: { status: "ERROR", message: "No device found for requested user" }, status: :not_found
         end
@@ -43,8 +49,10 @@ class DevicesController < ApplicationController
     def destroy
         device = Device.find_by(device_id: params[:id])
         if device
-            device.destroy
-            render json: { status: "SUCCESS", message: "Device deleted", data: device }, status: :ok
+            if check_current_user(device.user_id)
+                device.destroy
+                render json: { status: "SUCCESS", message: "Device deleted", data: device }, status: :ok
+            end
         else 
             render json: { status: "ERROR", message: "No device found for requested user" }, status: :not_found
         end
